requirements.txt
alembic==1.0.11
aniso8601==7.0.0
attrs==19.1.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
